requirements.txt
# Core APK Analysis
androguard==3.4.0a1
pyaxmlparser==0.3.31

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.26.0
pandas>=2.2.0
joblib>=1.3.0

# GUI Framework
# Tkinter comes with Python, no need to install via pip
pillow>=9.0.0

# Data Visualization
matplotlib>=3.8.0
seaborn>=0.12.0

# Threat Intelligence & Web Requests
requests>=2.31.0
urllib3>=2.0.0

# File Processing
python-magic>=0.4.27
# hashlib is built-in with Python, no pip install needed

# Additional Security Tools
yara-python>=4.3.1
pefile>=2023.2.7

# Development & Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0

# Performance
cython>=3.0.0

# Logging & Monitoring
coloredlogs>=15.0.1

# Configuration Management
pyyaml>=6.0.1

# Optional: Deep Learning
# tensorflow>=2.13.0
# torch>=2.0.0
